Reverts changes made to stripe_gateway_decorator.rb
